% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{gf_add_font}
\alias{gf_add_font}
\title{Add Google Fonts to 'tags' or 'tagList'}
\usage{
gf_add_font(tag_list = htmltools::tagList(), fontname = NULL,
  customstyle = character(), addstyle = TRUE)
}
\arguments{
\item{tag_list}{\code{htmltools::tags} or \code{htmltools::tagList}}

\item{fontname}{\code{character} valid Google Font name}

\item{customstyle}{\code{character} of custom weights and styles}

\item{addstyle}{\code{logical} to add style tag to use
the Google Font for body text}
}
\value{
\code{htmltools::tags} or \code{htmltools::tagList} with
          attached Google Font dependency
}
\description{
Add Google Fonts to 'tags' or 'tagList'
}

