% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{gf_font_dep}
\alias{gf_font_dep}
\title{Provide Dependency for Google Font}
\usage{
gf_font_dep(fontname = NULL, customstyle = character())
}
\arguments{
\item{fontname}{\code{character} valid Google Font name}

\item{customstyle}{\code{character} of custom weights and styles}
}
\value{
\code{htmltools::htmlDependency}
}
\description{
Provide Dependency for Google Font
}

