% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{gf_font_url}
\alias{gf_font_url}
\title{Generate URL for Google Font}
\usage{
gf_font_url(fontname = NULL, customstyle = character())
}
\arguments{
\item{fontname}{\code{character} valid Google Font name}

\item{customstyle}{\code{character} of custom weights and styles}
}
\value{
\code{character} url for Google Font
}
\description{
Generate URL for Google Font
}

