% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{gf_font_with_plus}
\alias{gf_font_with_plus}
\title{Substitute Plus for Space in Font Name}
\usage{
gf_font_with_plus(fontname = NULL)
}
\arguments{
\item{fontname}{\code{character} font name}
}
\value{
\code{character} font name with '+' instead of space
}
\description{
Substitute Plus for Space in Font Name
}

