% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{gf_font_with_space}
\alias{gf_font_with_space}
\title{Substitute Space for Plus in Font Name}
\usage{
gf_font_with_space(fontname = NULL)
}
\arguments{
\item{fontname}{\code{character} font name}
}
\value{
\code{character} font name with space instead of '+'
}
\description{
Substitute Space for Plus in Font Name
}

