% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{gf_add_font}
\alias{gf_add_font}
\title{Add Google Fonts to 'tags' or 'tagList'}
\usage{
gf_add_font(tags = htmltools::tagList(), fontname = NULL,
  customstyle = character())
}
\arguments{
\item{tags}{\code{htmltools::tags} or \code{htmltools::tagList}}

\item{fontname}{\code{character} valid Google Font name}

\item{customstyle}{\code{character} of custom weights and styles}
}
\value{
\code{htmltools::tags} or \code{htmltools::tagList} with
          attached Google Font dependency
}
\description{
Add Google Fonts to 'tags' or 'tagList'
}

